requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
Pillow==10.1.0
numpy==1.26.2
scikit-learn==1.3.2
gunicorn==21.2.0
python-dotenv==1.0.0

